create own canister components. refactor saving chats into component.
