+++ src/DeVinci_frontend/components/OwnCanister.svelte
@@ -1,0 +1,8 @@
+<script lang="ts">
+
+</script>
+
+<div>
+
+  hello
+</div>
